+++ mycodo/config.py
@@ -13,7 +13,7 @@ from flask_babel import lazy_gettext
 sys.path.append(os.path.abspath(os.path.dirname(__file__)))
 from config_translations import TRANSLATIONS
 
-MYCODO_VERSION = '8.10.0'
+MYCODO_VERSION = '8.10.1'
 ALEMBIC_VERSION = '4fedeb57e75b'
 
 #  FORCE_UPGRADE_MASTER